fix: 래플 서비스 버그 수정 및 기능 추가

+++ services/raffle-service/src/main/java/com/omc/raffle/infrastructure/client/dto/PreAuthRequest.java
@@ -1,8 +1,6 @@
 package com.omc.raffle.infrastructure.client.dto;
 
-import java.math.BigDecimal;
-
 public record PreAuthRequest(
         String billingKeyId,
-        BigDecimal amount
+        Long amount
 ) {}
